+++ CHANGES.rst
@@ -23,11 +23,15 @@ upgrading your version of coverage.py.
 Unreleased
 ----------
 
+- Fix: `issue 2138`_ describes a memory leak that happened when repeatedly
+  using the Coverage API with in-memory data. This is now fixed.
+
 - Fix: the markdown-formatted coverage report didn't fully escape special
   characters in file paths (`issue 2141`_). This would be very unlikely to
   cause a problem, but now it's done properly, thanks to `Ellie Ayla
   <pull 2142_>`_.
 
+.. _issue 2138: https://github.com/coveragepy/coveragepy/issues/2138
 .. _issue 2141: https://github.com/coveragepy/coveragepy/issues/2141
 .. _pull 2142: https://github.com/coveragepy/coveragepy/pull/2142